@@ -1,41 +1,42 @@
 {
-    "name": "frontend",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test"
-    },
-    "engines": {
-        "node": ">=18.19"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^18.0.3",
-        "@angular/common": "^18.0.3",
-        "@angular/compiler": "^18.0.3",
-        "@angular/core": "^18.0.3",
-        "@angular/forms": "^18.0.3",
-        "@angular/platform-browser": "^18.0.3",
-        "@angular/platform-browser-dynamic": "^18.0.3",
-        "@angular/router": "^18.0.3",
-        "rxjs": "~7.8.0",
-        "tslib": "^2.3.0",
-        "zone.js": "~0.14.7"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^18.0.4",
-        "@angular/cli": "~18.0.4",
-        "@angular/compiler-cli": "^18.0.3",
-        "@types/jasmine": "~4.3.0",
-        "jasmine-core": "~4.6.0",
-        "karma": "~6.4.0",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.0",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.0.0",
-        "typescript": "~5.4.5"
-    }
+  "name": "frontend",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "engines": {
+    "node": ">=18.19"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^20.0.1",
+    "@angular/common": "^20.0.1",
+    "@angular/compiler": "^20.0.1",
+    "@angular/core": "^20.0.1",
+    "@angular/forms": "^20.0.1",
+    "@angular/platform-browser": "^20.0.1",
+    "@angular/platform-browser-dynamic": "^20.0.1",
+    "@angular/router": "^20.0.1",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.1"
+  },
+  "devDependencies": {
+    "@angular/build": "^20.0.1",
+    "@angular/cli": "~20.0.1",
+    "@angular/compiler-cli": "^20.0.1",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "less": "^4.2.0",
+    "typescript": "~5.8.3"
+  }
 }
