@@ -2,8 +2,8 @@ import { Component } from "@angular/core";
 
 @Component({
     selector: 'fsss-footer',
-    templateUrl: './footer.component.html',
-    styleUrls: ['./footer.component.less'],
+    templateUrl: 'footer.component.html',
+    styleUrl: 'footer.component.less',
 })
 export class FooterComponent {
 
