@@ -1,9 +1,11 @@
 import { Component } from "@angular/core";
+import {RouterModule} from "@angular/router";
 
 @Component({
     selector: 'fsss-header',
-    templateUrl: './header.component.html',
-    styleUrls: ['./header.component.less'],
+    templateUrl: 'header.component.html',
+    styleUrl: 'header.component.less',
+    imports: [RouterModule],
 })
 export class HeaderComponent {
 
