@@ -1,8 +1,8 @@
 import { Component } from "@angular/core";
 
 @Component({
-    templateUrl: './about.view.html',
-    styleUrls: ['./about.view.less'],
+    templateUrl: 'about.view.html',
+    styleUrl: 'about.view.less',
 })
 export class AboutView {
 
